# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SnackVeda e-commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new B2C or B2B user
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthUser" }
        "400":
          description: Invalid
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Log in
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginBody" }
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthUser" }
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /account/addresses:
    get:
      operationId: listMyAddresses
      tags: [account]
      summary: List my saved addresses
      responses:
        "200":
          description: Addresses
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Address" }
    post:
      operationId: createMyAddress
      tags: [account]
      summary: Save a new shipping address
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateAddressBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Address" }
...
  /admin/dashboard:
    get:
      operationId: getAdminDashboard
      tags: [admin]
      summary: Admin KPI dashboard
      responses:
        "200":
          description: Dashboard
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminDashboard" }
        "403":
          description: Forbidden
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /admin/products:
    get:
      operationId: listAdminProducts
      tags: [admin]
      summary: List all products (admin)
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Product" }
    post:
      operationId: createAdminProduct
      tags: [admin]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateProductBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
  /admin/products/{id}:
    patch:
      operationId: updateAdminProduct
      tags: [admin]
      summary: Update a product (stock, status, prices, etc.)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateProductBody" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
  /admin/customers:
    get:
      operationId: listAdminCustomers
      tags: [admin]
      summary: List customers
      parameters:
        - in: query
          name: type
          required: false
          schema:
            type: string
            enum: [b2c, b2b]
      responses:
        "200":
          description: Customers
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AdminCustomer" }
...
  /admin/orders:
    get:
      operationId: listAdminOrders
      tags: [admin]
      summary: List all orders (admin)
      parameters:
        - in: query
          name: orderType
          required: false
          schema: { $ref: "#/components/schemas/OrderType" }
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Order" }
  /admin/orders/{id}/status:
    patch:
      operationId: updateAdminOrderStatus
      tags: [admin]
      summary: Update order status
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateOrderStatusBody" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
...
  /admin/payments/{orderId}/confirm:
    patch:
      operationId: confirmAdminPayment
      tags: [admin]
      summary: Mark a payment as received (creates invoice + confirms order)
      parameters:
        - in: path
          name: orderId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ConfirmPaymentBody" }
      responses:
        "200":
          description: Confirmed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]
    UserRole:
      type: string
      enum: [b2c_customer, b2b_customer, super_admin]
    CustomerType:
      type: string
      enum: [retail, kirana, modern_retail, gym, pharmacy, cafe, corporate]
    B2bStatus:
      type: string
      enum: [pending, approved, rejected]
    ProductCategory:
      type: string
      enum: [healthy_chips, makhana, superpuffs]
    ProductStatus:
      type: string
      enum: [active, inactive, out_of_stock]
    OrderType:
      type: string
      enum: [b2c, b2b]
    OrderStatus:
      type: string
      enum: [pending, confirmed, dispatched, delivered, cancelled]
    PaymentMethod:
      type: string
      enum: [upi, bank_transfer, payment_link]
    PaymentStatus:
      type: string
      enum: [pending, received, failed, refunded]
    AuthUser:
      type: object
      properties:
        id: { type: string }
        email: { type: string }
        fullName: { type: string }
        role: { $ref: "#/components/schemas/UserRole" }
        phone: { type: ["string", "null"] }
        b2bStatus:
          oneOf:
            - { $ref: "#/components/schemas/B2bStatus" }
            - { type: "null" }
      required: [id, email, fullName, role, phone, b2bStatus]
...
    RegisterBody:
      type: object
      properties:
        email: { type: string }
        password: { type: string, minLength: 6 }
        fullName: { type: string }
        phone: { type: ["string", "null"] }
        accountType:
          type: string
          enum: [b2c, b2b]
        businessName: { type: ["string", "null"] }
        businessType:
          oneOf:
            - { $ref: "#/components/schemas/CustomerType" }
            - { type: "null" }
        gstNumber: { type: ["string", "null"] }
        businessAddress: { type: ["string", "null"] }
      required: [email, password, fullName, accountType]
    LoginBody:
      type: object
      properties:
        email: { type: string }
        password: { type: string }
      required: [email, password]
...
    Address:
      type: object
      properties:
        id: { type: string }
        fullName: { type: string }
        phone: { type: string }
        line1: { type: string }
        line2: { type: ["string", "null"] }
        city: { type: string }
        state: { type: string }
        pincode: { type: string }
        isDefault: { type: boolean }
      required: [id, fullName, phone, line1, line2, city, state, pincode, isDefault]
    CreateAddressBody:
      type: object
      properties:
        fullName: { type: string }
        phone: { type: string }
        line1: { type: string }
        line2: { type: ["string", "null"] }
        city: { type: string }
        state: { type: string }
        pincode: { type: string }
      required: [fullName, phone, line1, city, state, pincode]
    Product:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        category: { $ref: "#/components/schemas/ProductCategory" }
        variant: { type: ["string", "null"] }
        b2cPrice: { type: number }
        b2bPrice: { type: number }
        moq: { type: integer }
        cartonQty: { type: integer }
        gstPercent: { type: number }
        hsnCode: { type: string }
        shelfLifeMonths: { type: integer }
        weightGrams: { type: integer }
        description: { type: ["string", "null"] }
        stockQty: { type: integer }
        status: { $ref: "#/components/schemas/ProductStatus" }
        sortOrder: { type: integer }
        imageUrl: { type: ["string", "null"] }
      required: [id, name, slug, category, variant, b2cPrice, b2bPrice, moq, cartonQty, gstPercent, hsnCode, shelfLifeMonths, weightGrams, description, stockQty, status, sortOrder, imageUrl]
...
    CreateProductBody:
      type: object
      properties:
        name: { type: string }
        slug: { type: string }
        category: { $ref: "#/components/schemas/ProductCategory" }
        variant: { type: ["string", "null"] }
        b2cPrice: { type: number }
        b2bPrice: { type: number }
        moq: { type: integer }
        cartonQty: { type: integer }
        gstPercent: { type: number }
        hsnCode: { type: string }
        shelfLifeMonths: { type: integer }
        weightGrams: { type: integer }
        description: { type: ["string", "null"] }
        stockQty: { type: integer }
      required: [name, slug, category, b2cPrice, b2bPrice, moq, cartonQty, gstPercent, hsnCode, shelfLifeMonths, weightGrams, stockQty]
    UpdateProductBody:
      type: object
      properties:
        name: { type: ["string", "null"] }
        b2cPrice: { type: ["number", "null"] }
        b2bPrice: { type: ["number", "null"] }
        stockQty: { type: ["integer", "null"] }
        status:
          oneOf:
            - { $ref: "#/components/schemas/ProductStatus" }
            - { type: "null" }
        description: { type: ["string", "null"] }
...
    OrderItemView:
      type: object
      properties:
        id: { type: string }
        productId: { type: string }
        productName: { type: string }
        slug: { type: string }
        category: { $ref: "#/components/schemas/ProductCategory" }
        quantity: { type: integer }
        unitPrice: { type: number }
        gstPercent: { type: number }
        gstAmount: { type: number }
        lineTotal: { type: number }
        hsnCode: { type: string }
      required: [id, productId, productName, slug, category, quantity, unitPrice, gstPercent, gstAmount, lineTotal, hsnCode]
    Order:
      type: object
      properties:
        id: { type: string }
        orderNumber: { type: string }
        userId: { type: string }
        customerName: { type: string }
        customerEmail: { type: string }
        orderType: { $ref: "#/components/schemas/OrderType" }
        status: { $ref: "#/components/schemas/OrderStatus" }
        subtotal: { type: number }
        discountAmount: { type: number }
        discountPercent: { type: number }
        gstAmount: { type: number }
        shippingCharge: { type: number }
        totalAmount: { type: number }
        notes: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItemView" }
        shippingAddress:
          oneOf:
            - { $ref: "#/components/schemas/Address" }
            - { type: "null" }
        payment:
          oneOf:
            - { $ref: "#/components/schemas/PaymentView" }
            - { type: "null" }
        invoiceNumber: { type: ["string", "null"] }
      required: [id, orderNumber, userId, customerName, customerEmail, orderType, status, subtotal, discountAmount, discountPercent, gstAmount, shippingCharge, totalAmount, notes, createdAt, items, shippingAddress, payment, invoiceNumber]
    PaymentView:
      type: object
      properties:
        id: { type: string }
        paymentMethod: { $ref: "#/components/schemas/PaymentMethod" }
        paymentStatus: { $ref: "#/components/schemas/PaymentStatus" }
        amount: { type: number }
        referenceNumber: { type: ["string", "null"] }
        paidAt: { type: ["string", "null"], format: date-time }
      required: [id, paymentMethod, paymentStatus, amount, referenceNumber, paidAt]
...
    AdminDashboard:
      type: object
      properties:
        todayOrdersCount: { type: integer }
        pendingPaymentsCount: { type: integer }
        thisMonthRevenue: { type: number }
        lowStockCount: { type: integer }
        totalCustomers: { type: integer }
        pendingB2bApplications: { type: integer }
        recentOrders:
          type: array
          items: { $ref: "#/components/schemas/Order" }
        ordersByCategory:
          type: array
          items:
            type: object
            properties:
              category: { $ref: "#/components/schemas/ProductCategory" }
              quantity: { type: integer }
              revenue: { type: number }
            required: [category, quantity, revenue]
      required: [todayOrdersCount, pendingPaymentsCount, thisMonthRevenue, lowStockCount, totalCustomers, pendingB2bApplications, recentOrders, ordersByCategory]
    AdminCustomer:
      type: object
      properties:
        id: { type: string }
        email: { type: string }
        fullName: { type: string }
        phone: { type: ["string", "null"] }
        role: { $ref: "#/components/schemas/UserRole" }
        customerType:
          oneOf:
            - { $ref: "#/components/schemas/CustomerType" }
            - { type: "null" }
        businessName: { type: ["string", "null"] }
        gstNumber: { type: ["string", "null"] }
        businessAddress: { type: ["string", "null"] }
        b2bStatus:
          oneOf:
            - { $ref: "#/components/schemas/B2bStatus" }
            - { type: "null" }
        ordersCount: { type: integer }
        createdAt: { type: string, format: date-time }
      required: [id, email, fullName, phone, role, customerType, businessName, gstNumber, businessAddress, b2bStatus, ordersCount, createdAt]
...
    UpdateOrderStatusBody:
      type: object
      properties:
        status: { $ref: "#/components/schemas/OrderStatus" }
      required: [status]
...
    ConfirmPaymentBody:
      type: object
      properties:
        referenceNumber: { type: string }
      required: [referenceNumber]