# lib/api-zod/src/generated/types/adminCustomer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SnackVeda e-commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { B2bStatus } from "./b2bStatus";
import type { CustomerType } from "./customerType";
import type { UserRole } from "./userRole";

export interface AdminCustomer {
  id: string;
  email: string;
  fullName: string;
  /** @nullable */
  phone: string | null;
  role: UserRole;
  customerType: CustomerType | null;
  /** @nullable */
  businessName: string | null;
  /** @nullable */
  gstNumber: string | null;
  /** @nullable */
  businessAddress: string | null;
  b2bStatus: B2bStatus | null;
  ordersCount: number;
  createdAt: Date;
}