# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SnackVeda e-commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { Address } from "./address";
import type { OrderItemView } from "./orderItemView";
import type { OrderStatus } from "./orderStatus";
import type { OrderType } from "./orderType";
import type { PaymentView } from "./paymentView";

export interface Order {
  id: string;
  orderNumber: string;
  userId: string;
  customerName: string;
  customerEmail: string;
  orderType: OrderType;
  status: OrderStatus;
  subtotal: number;
  discountAmount: number;
  discountPercent: number;
  gstAmount: number;
  shippingCharge: number;
  totalAmount: number;
  /** @nullable */
  notes: string | null;
  createdAt: Date;
  items: OrderItemView[];
  shippingAddress: Address | null;
  payment: PaymentView | null;
  /** @nullable */
  invoiceNumber: string | null;
}