# lib/api-zod/src/generated/types/paymentMethod.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SnackVeda e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export type PaymentMethod = (typeof PaymentMethod)[keyof typeof PaymentMethod];

export const PaymentMethod = {
  upi: "upi",
  bank_transfer: "bank_transfer",
  payment_link: "payment_link",
} as const;