# lib/api-zod/src/generated/types/userRole.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SnackVeda e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  b2c_customer: "b2c_customer",
  b2b_customer: "b2b_customer",
  super_admin: "super_admin",
} as const;